backward compatible regexp based callbacks & inlines
